+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Gives user feedback on how many chapter titles were successfully uploaded
+
+### Fixed
+- Better display for error messages and upload progress
+
 ## [1.1.0] - 2026-2-16
 ### Added
 - Alert if no chapter titles would be uploaded due to the offset